uriparse

Module scheme

Scheme Component

See [RFC3986, Section 3.5]. For a list of the listed schemes, see iana.org.

Structs

UnregisteredScheme

A scheme that is not in the registered schemes.

UnregisteredSchemeError

An error representing that the unregistered scheme was an invalid scheme, or it was actually a registered scheme.

Enums

Scheme

The scheme component as defined in [RFC3986, Section 3.5]. The schemes listed here come from iana.org. Any scheme not listed there is considered unregistered and will be contained in [Scheme::UnregisteredScheme].

SchemeError

An error representing an invalid scheme.

SchemeStatus

The registration status of a scheme. See RFC 7595 for more information.