uriparse

Module uri

URIs, Relative References, and URI References

See RFC3986.

This module is composed of three primary types URI, [RelativeReference], and URIReference that are all very similar. The first thing to note is that URIs and relative references are types of URI references. They differ in only one way: URIs have schemes, while relative references do not.

As a result, choose the type that best fits your use case. If you need absolute URIs, you should use URI, but if you want relative references (e.g. "/" in a GET request) use [RelativeReference]. If you can accept both, then use URIReference.

Structs

URI

A Uniform Resource Identifier (URI) as defined in RFC3986.

URIBuilder

A builder type for [URI].

Enums

URIError

An error representing an invalid URI.